smallfix before fixing structure

+++ client/src/pages/DeleteRecipe.jsx
@@ -1,18 +1,20 @@
 import React from 'react';
 import axios from 'axios';
 
-const DeleteRecipe = ({ recipeID, userID, authToken }) => {
+const DeleteRecipe = ({ recipeID, userID, authToken, updateRecipes }) => {
   const handleDelete = async () => {
-    try {
-      await axios.delete(`http://localhost:3001/delete-recipe/${recipeID}`, {
-        data: { userID },
-        headers: { authorization: authToken } // Include the authentication token in headers
-      });
-      alert("Recipe deleted successfully");
-      // You can redirect the user to another page or update the state as needed
-    } catch (error) {
-      console.error(error);
-      alert("Failed to delete recipe");
+    if (window.confirm("Are you sure you want to delete this recipe?")) {
+      try {
+        await axios.delete(`http://localhost:3001/delete-recipe/${recipeID}`, {
+          data: { userID },
+          headers: { authorization: authToken }
+        });
+        alert("Recipe deleted successfully");
+        updateRecipes(); // Refresh recipes after deletion
+      } catch (error) {
+        console.error(error);
+        alert("Failed to delete recipe");
+      }
     }
   };
